% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{break_function}
\alias{break_function}
\title{break_function}
\usage{
break_function(func, check_output = FALSE)
}
\arguments{
\item{func}{a function to be broken down}

\item{check_output}{should we write the list out}
}
\value{
a string that will run in an R sesseion to
1. Define the function,
2. parse arguments to the function,
3. return the outputs of t
}
\description{
builds callable structure from the function body,
}
\author{
Joe Peskett
}
