% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_from_function}
\alias{component_from_function}
\title{component_from_function}
\usage{
component_from_function(func, base_image, component_output_file = NULL)
}
\arguments{
\item{func}{an R function}

\item{base_image}{a base image to use for the component}

\item{component_output_file}{an optional argument to save the component to a YAML file}
}
\description{
this function builds a Kubeflow Pipelines component from an R function.
Note that any packages required by the function, should be added into the
}
\details{
Components are built up of inputs and outputs which are used as argumens to an implementation of a container.
}
\author{
Joe Peskett
}
