% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{function_to_interface}
\alias{function_to_interface}
\title{function_to_interface}
\usage{
function_to_interface(func, check_output = FALSE)
}
\arguments{
\item{func}{The function to break up}

\item{check_output}{should we save the output; for debugging purposes}
}
\description{
given a function, return it as a string
and create an interface that will take arguments when called from
terminal
}
\author{
Joe Peskett
}
