% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubeflowTool.R
\docType{package}
\name{kfTool-package}
\alias{kfTool}
\alias{kfTool-package}
\title{kfTool: A basic package for building dockerfiles and Kubeflow pipeline component in R, with R.}
\description{
kfTool is for building R components from R functions.
}
\keyword{internal}
