% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{load_component_from_file}
\alias{load_component_from_file}
\title{load_component_from_file}
\usage{
load_component_from_file(component_path)
}
\arguments{
\item{component_path}{location of the component to load}
}
\description{
load a kubeflow pipelines component from file
}
\author{
Joe Peskett
}
