% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_tools.R
\name{step_CMD}
\alias{step_CMD}
\title{step_CMD}
\usage{
step_CMD(dockerfile, args)
}
\arguments{
\item{dockerfile}{A dockerfile - note that this will not check to see if you have a valid dockerfile!}

\item{args}{A list of arguments to use as part of your CMD step}
}
\description{
add a CMD step in your dockerfile
}
\author{
Joe Peskett
}
