% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_tools.R
\name{step_COPY}
\alias{step_COPY}
\title{step_COPY}
\usage{
step_COPY(dockerfile, local_path, dockerfile_path)
}
\arguments{
\item{dockerfile}{A dockerfile - note that this will not check to see if you have a valid dockerfile!}

\item{local_path}{File path to copy}

\item{dockerfile_path}{Where to copy to in the dockerfile}
}
\description{
add COPY step
}
\author{
Joe Peskett
}
