% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_tools.R
\name{step_CUSTOM}
\alias{step_CUSTOM}
\title{step_CUSTOM}
\usage{
step_CUSTOM(dockerfile, command, args)
}
\arguments{
\item{dockerfile}{A dockerfile - note that this will not check to see if you have a valid dockerfile!}

\item{command}{Command to execute}

\item{args}{These should be formatted in a list}
}
\description{
add CUSTOM step if commands are not covered in this package
}
\author{
Joe Peskett
}
