% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_tools.R
\name{step_FROM}
\alias{step_FROM}
\title{step_FROM}
\usage{
step_FROM(dockerfile, image_name, image_tag = "latest")
}
\arguments{
\item{dockerfile}{A dockerfile - note that this will not check to see if you have a valid dockerfile!}

\item{image_name}{The name of the image to build from}

\item{image_tag}{The tag of the image. Defaults to latest}
}
\description{
add a FROM step, usually used for selecting a base image
}
\author{
Joe Peskett
}
