% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_tools.R
\name{step_RUN}
\alias{step_RUN}
\title{step_RUN}
\usage{
step_RUN(dockerfile, args)
}
\arguments{
\item{dockerfile}{A dockerfile - note that this will not check to see if you have a valid dockerfile!}

\item{args}{A list of arguments}
}
\description{
add a RUN step in your dockerfile
}
\author{
Joe Peskett
}
