% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{string_to_func}
\alias{string_to_func}
\title{string_to_func}
\usage{
string_to_func(func_as_string)
}
\arguments{
\item{func_as_string}{this will be the output of break_funcion}
}
\description{
take a string and evaluate it to produce a function called 'worker'
}
\author{
Joe Peskett
}
