% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_tools.R
\name{writeDockerfile}
\alias{writeDockerfile}
\title{writeDockerfile}
\usage{
writeDockerfile(dockerfile, file)
}
\arguments{
\item{dockerfile}{final dockerfile}

\item{file}{where to write the file}
}
\description{
Writes the dockerfile to file
}
\author{
Joe Peskett
}
